Project update 12:28 PM 5-16-2016
Made edits to the win, main, gameover and game scenes. Everything in the
win, main and gameover scenes are static. Game scene edits: main camera,
directional light, floor, walls, base, and ui buttons are static. player
has the player tag. Created an empty game object to the canvas and
applied the game controller script to it. Added ui elements to the
script in the inspector menu.